+++ workers/ohsome_quality_analyst/definitions.py
@@ -76,7 +76,7 @@ INDICATOR_LAYER = (
     ("MappingSaturation", "mapaction_rail_length"),
     ("MappingSaturation", "mapaction_lakes_area"),
     ("MappingSaturation", "mapaction_rivers_length"),
-    ("MappingSaturation", "basic_infrastructure"),
+    ("MappingSaturation", "infrastructure_lines"),
     ("MappingSaturation", "poi"),
     ("MappingSaturation", "lulc"),
     ("Currentness", "major_roads_count"),