+++ CHANGELOG.md
@@ -1,6 +1,23 @@
 # Changelog
 
+
 ## Current Main
+
+## Breaking Changes
+
+- Rename layer `ideal_vgi_infrastructure` to `infrastructure_lines` ([#416])
+
+### New Features
+
+- Improve Currentness indicator ([#274])
+- Add new report `MultilevelCurrentness` ([#403])
+
+[#274]: https://github.com/GIScience/ohsome-quality-analyst/pull/274
+[#403]: https://github.com/GIScience/ohsome-quality-analyst/pull/403
+[#416]: https://github.com/GIScience/ohsome-quality-analyst/pull/416
+
+
+## 0.11.0
 
 ### Breaking Changes
 
@@ -19,10 +36,9 @@
 
 - Add new representative report `RoadReport` ([#357])
 - Add new representative report `BuildingReport` ([#356])
-- Add ratio_filter to `building_count` layer ([#356])
+- Add ratio filter to `building_count` layer ([#356])
 - Configure OQT using files or environment variables ([#255])
 - Generalize result values of all Indicators by introducing a result class value ([#369])
-- Add new report 'MultilevelCurrentness' ([#403])
 
 ### Other Changes
 
@@ -30,6 +46,7 @@
 - Substitute Simple Report with a Report named Minimal for testing purposes ([#342] [#385])
 - Add a minimal Indicator for testing purposes ([#383])
 - Remove database scripts ([#392])
+- fix: building completeness figure  ([#410])
 
 ### How to Upgrade
 
@@ -38,7 +55,6 @@
 - To continue to retrieve additional data of an Indicator or Report provided in an API response, you need to set the API request parameter `include_data` to `True` ([#370])
 - Rename environment variable `OHSOME_API`  `OQT_OHSOME_API` ([#255])
 - Make sure to rename the API query parameter `layerName` to `layerKey` and API endpoint `listLayerNames` to `listLayerKeys` ([#376])
-- To continue to retrieve the properties of the GeoJSON API response as flat list, you need to set the API request parameter `flatten` to `True` ([#375])
 - If you run your own database, please delete the result table before upgrading ([#369])
 - Rename endpoints ([#397]):
 
@@ -66,8 +82,7 @@
 [#392]: https://github.com/GIScience/ohsome-quality-analyst/pull/392
 [#396]: https://github.com/GIScience/ohsome-quality-analyst/pull/396
 [#397]: https://github.com/GIScience/ohsome-quality-analyst/pull/397
-[#403]: https://github.com/GIScience/ohsome-quality-analyst/pull/403
-
+[#410]: https://github.com/GIScience/ohsome-quality-analyst/pull/410
 
 ## 0.10.1